requirements.txt
py-mmdb-encoder==1.0.4
maxminddb==1.5.4
netaddr==0.7.20
requests